Storing images in firebase
It is not ok to have images disappear and be replaced by
server-side images, which need a reload, once we have uploaded them.

Approach:

When the user adds an image, we have a handle on a File object in
JavaScript.

Assign it a name based on the current bake ID and a random string.

Assign it an extension based on the mime-type (not implemented yet, only
PNG supported)

Get a blob:// URL from the File and write it into a database field
called :cache under the filename of the image.

Save the image step in the databse including a filename.

Re-render the image using the data from the cache

Send the image off to firebase to be saved.

Next: on loading an existing bake, load the images from firebase and on
each successful load, put the image src in the cache so the browser can
do the right thing.