Rehydration!
We can now load a bake BACK from Firestore.

We need to map FS's weird JS back to proper CLJS structures, which is a
bit of a pain.

Should make the fs IO event handlers a bit more generic, maybe provide
some nicer way of hooking into them and then put them in the firebase
namespace.

Also need to load the photos and put their urls in cache:. Probably not
a well-named key.

Also need to be able to abort/declare a bake finished.